# backend/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SkillBridge API Ecosystem",
    "description": "Skill Intelligence and Career Readiness Platform API Specifications",
    "contact": {
      "name": "SkillBridge Development Team",
      "email": "support@skillbridge.internal"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://springdoc.org"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/projects/{projectId}/progress": {
      "put": {
        "tags": [
          "Project Service"
        ],
        "summary": "Update project completion progress",
        "operationId": "updateProgress",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectProgressRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseProjectProgress"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning/{courseId}/progress": {
      "post": {
        "tags": [
          "Learning Recommendation Service"
        ],
        "summary": "Update learning progress for a course",
        "operationId": "updateProgress_1",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLearningProgress"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence": {
      "get": {
        "tags": [
          "Evidence & Verification Service"
        ],
        "summary": "Get all evidence items for authenticated student",
        "operationId": "getEvidence_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListEvidence"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Evidence & Verification Service"
        ],
        "summary": "Record a new evidence item",
        "operationId": "addEvidence",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEvidenceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEvidence"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/learning/{courseId}": {
      "get": {
        "tags": [
          "Learning Recommendation Service"
        ],
        "summary": "Get course details by ID",
        "operationId": "getCourseById",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseCourse"
                }
              }
            }
          }
        }
      }
    },
    "/api/learning/skill/{skillId}": {
      "get": {
        "tags": [
          "Learning Recommendation Service"
        ],
        "summary": "Get courses for a specific skill",
        "operationId": "getCoursesBySkillId",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListCourse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{jobId}": {
      "get": {
        "tags": [
          "Job Service"
        ],
        "summary": "Get job details by ID",
        "operationId": "getJobById",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJob"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/{jobId}/match": {
      "get": {
        "tags": [
          "Job Service"
        ],
        "summary": "Calculate student competency match percentage for a job posting",
        "operationId": "getJobMatch",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJobMatchDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/{id}": {
      "get": {
        "tags": [
          "Evidence & Verification Service"
        ],
        "summary": "Get specific evidence by ID",
        "operationId": "getEvidenceById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseEvidence"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Evidence & Verification Service"
        ],
        "summary": "Delete an evidence item",
        "operationId": "deleteEvidence_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/skill/{skillId}": {
      "get": {
        "tags": [
          "Evidence & Verification Service"
        ],
        "summary": "Get evidence items for a specific skill",
        "operationId": "getEvidenceBySkillId",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListEvidence"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile/evidence/{id}": {
      "delete": {
        "tags": [
          "Profile Service"
        ],
        "summary": "Delete an evidence item",
        "operationId": "deleteEvidence",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseString"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProjectProgressRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "SUBMITTED",
              "EVALUATED"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "githubRepoUrl": {
            "type": "string"
          }
        }
      },
      "ApiResponseProjectProgress": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/ProjectProgress"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "ProjectProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "projectId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "SUBMITTED",
              "EVALUATED"
            ]
          },
          "githubRepoUrl": {
            "type": "string"
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "submissionNotes": {
            "type": "string"
          },
          "evaluationScore": {
            "type": "integer",
            "format": "int32"
          },
          "feedback": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Evidence": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "RESUME",
              "CERTIFICATE",
              "PROJECT",
              "ASSESSMENT",
              "INTERNSHIP",
              "PRACTICAL_TASK"
            ]
          },
          "fileUrl": {
            "type": "string"
          },
          "verifiedAt": {
            "type": "string"
          },
          "score": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponseString": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseEvidence": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Evidence"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "ProgressUpdateRequest": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ApiResponseLearningProgress": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/LearningProgress"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "LearningProgress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          },
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "progressPercent": {
            "type": "integer",
            "format": "int32"
          },
          "lastAccessedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateEvidenceRequest": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "RESUME",
              "CERTIFICATE",
              "PROJECT",
              "ASSESSMENT",
              "INTERNSHIP",
              "PRACTICAL_TASK"
            ]
          },
          "fileUrl": {
            "type": "string"
          },
          "score": {
            "type": "string"
          }
        }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "difficulty": {
            "type": "string"
          },
          "skillsCovered": {
            "type": "string"
          },
          "competenciesCovered": {
            "type": "string"
          },
          "isFree": {
            "type": "boolean"
          },
          "recommendationReason": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "skillId": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListEvidence": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Evidence"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseListCourse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "ApiResponseCourse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Course"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "jobTitle": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "employmentType": {
            "type": "string"
          },
          "salary": {
            "type": "string"
          },
          "requiredSkills": {
            "type": "string"
          },
          "applicationUrl": {
            "type": "string"
          },
          "remote": {
            "type": "boolean"
          },
          "experienceLevel": {
            "type": "string"
          }
        }
      },
      "ApiResponseJob": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/Job"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "ApiResponseJobMatchDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/JobMatchDto"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string"
          }
        }
      },
      "JobMatchDto": {
        "type": "object",
        "properties": {
          "job": {
            "$ref": "#/components/schemas/Job"
          },
          "competencyMatchPercentage": {
            "type": "integer",
            "format": "int32"
          },
          "matchingSkills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "missingSkills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "matchCategory": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}